// interfaces/StyleInterface.ts
import { TextStyle, ViewStyle } from "react-native";

export type ForgotPasswordStyles = {
  container: ViewStyle;
  logo: TextStyle;
  formContainer: ViewStyle;
  title: TextStyle;
  input: TextStyle;
  error: TextStyle;
  link: TextStyle;
};

export type SignupStyles = {
  container: ViewStyle;
  logo: TextStyle;
  formContainer: ViewStyle;
  title: TextStyle;
  input: TextStyle;
  error: TextStyle;
  link: TextStyle;
};

export type LoginStyles = {
  container: ViewStyle;
  logo: TextStyle;
  formContainer: ViewStyle;
  title: TextStyle;
  input: TextStyle;
  error: TextStyle;
  link: TextStyle;
  linksContainer: ViewStyle;
};